long commit, not much hate tho.
Created PersistentStorage (maybe not its final name), a module to persistently storage data. A whole bunch of this is discussed in the trello page, but its no harm to use it here.

persistent.json has some basic config options as it was going to be the first module to use persistent storage. Ana is a simpler one and thus the first one to get changes. It already has random boobs!!

anyhow. My biggest problem is that now json handles a dictionary and I have to either learn how to automatically navigate a dictionary or learn how to serialize a pojo stile object.

Other concern is to store the data back. whitch may push a mongo server earlier into production. it's going to be so cool ;). Don't worry, mongo servers are light, but it may be fun to deal with several aplications hitting the same server and expecting different information.

there is a way to pass an argument to a function with a default set of data, there is an other way to say that some argument wont be passed leaving it to allways default. If it's a big argument (not a basic type) it wont create it again and again, like a big hoarding data argument that python won't creat again and again but rather reuse it.

I should get the telegram api reference for offline access